RectangleB first draft, need to think about edge cases -possibly change the setters

+++ test/Einstein.java
@@ -15,10 +15,10 @@ public class Einstein
      */
     public static void main(String[]args)
     {
-        int reversedNum, hundreds, tens, units, diff, reversedDiff;
+        long reversedNum, hundreds, tens, units, diff, reversedDiff;
         System.out.println("Welcome to the Einstein magic game. \nPlease enter a 3 digit positive number whose first and last digits are different:");
         Scanner scan = new Scanner (System.in);
-        long num = scan.nextInt();
+        long num = scan.nextLong();
         // Checks if the input is a 3 digit positive number
         if ((num > 999) || (num < 100))
             System.out.println("The number you entered is not a 3 digit positive number");
